requirements.txt
aiohttp==3.8.0
aiosignal==1.2.0
async-timeout==4.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.14.5
chardet==4.0.0
charset-normalizer==2.0.7
constantly==15.1.0
cryptography==3.4.7
cycler==0.10.0
Cython==0.29.23
dateparser==1.1.0
frozenlist==1.2.0
greenlet==1.1.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
kiwisolver==1.3.2
matplotlib==3.4.3
multidict==5.2.0
numpy==1.21.4
pandas==1.3.4
Pillow==8.3.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyMySQL==1.0.2
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-binance==1.0.15
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
regex==2021.11.10
requests==2.26.0
service-identity==18.1.0
six==1.16.0
SQLAlchemy==1.4.26
Twisted==21.2.0
txaio==21.2.1
typing-extensions==3.10.0.2
tzdata==2021.5
tzlocal==4.1
ujson==4.2.0
urllib3==1.26.7
websockets==9.1
yarl==1.7.2
zope.interface==5.4.0
